backend/requirements.txt
# FastAPI
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings

# Database
sqlalchemy[asyncio]
asyncpg
alembic

# Auth
python-jose[cryptography]
bcrypt==4.0.1 
passlib==1.7.4
# Redis
redis

# LLM Providers
langchain
langchain-core
langchain-community
langchain-groq
langchain-openai
langchain-anthropic
langchain-google-genai
langgraph
psycopg2-binary asyncpg

# RAG
pinecone
langchain-pinecone
pypdf
pandas
python-docx
python-pptx
openpyxl
xlrd

# File Storage
cloudinary

# Audio
groq
openai
aiohttp

# Observability
langsmith

# Utilities
python-dotenv
httpx
aiofiles